refactor: split aum_tools.py into slmkiii/aum/ subpackage
Five focused modules instead of one 525-line top-level file:
  - slmkiii/aum/codec.py     — AumMsgType IntEnum + MSG_TYPE_* aliases
  - slmkiii/aum/archiver.py  — NSKeyedArchiver decode/encode + ArchiverBuilder
  - slmkiii/aum/midimap.py   — AumMidiMapping + read/write_aum_midimap
  - slmkiii/aum/session.py   — AumPlugin/AumChannel/AumSession + read_aum_session
  - slmkiii/aum/cli.py       — inspect-mapping / inspect-session CLI

slmkiii/aum/__init__.py re-exports the public API.

Updates:
  - All callers (slmkiii/harvest.py, tests/test_aum_tools.py,
    scripts/generate_aum_mappings.py) now import from slmkiii.aum.
  - Deleted top-level aum_tools.py.
  - Added slmkiii-aum console script entry point.
  - Drop pyyaml dep (only used by deleted controlmap.spec_loader).

Tests: 109 passing.
