+++ pyproject.toml
@@ -18,7 +18,6 @@ classifiers = [
 dependencies = [
     "mido>=1.3.3",
     "python-rtmidi>=1.5.8",
-    "pyyaml>=6.0.3",
 ]
 
 [project.optional-dependencies]
@@ -27,6 +26,7 @@ dev = ["bpython>=0.26", "coverage"]
 
 [project.scripts]
 slmkiii = "slmkiii.cli:main"
+slmkiii-aum = "slmkiii.aum.cli:main"
 
 [project.urls]
 Repository = "https://github.com/TheFermiSea/slmkiii"
